feat: Add lazy loading for images and update user dashboard layout
This commit adds lazy loading for images to improve page load performance. It also updates the layout for the user dashboard.

chore: Update npm dependencies and add Zustand library

This commit updates the `package.json` and `package-lock.json` files to include the `zustand` library as a new dependency. The version added is `5.0.0-rc.2`.

feat: Update login page and integrate API for profile and homepage user

This commit updates the login page and integrates the API for the profile page and homepage user.

chore: Update eslint configuration

Update the .eslintrc.json file to include additional rules for TypeScript. Disable the "@typescript-eslint/no-empty-object-type" and "@typescript-eslint/no-unused-vars" rules.

chore: Update npm dependencies to include @tanstack/react-query

This commit updates the `package.json` and `package-lock.json` files to include the `@tanstack/react-query` library as a new dependency. The version added is `5.56.2`.